+++ test/equality.mjs
@@ -5,40 +5,40 @@ group('Equality tests', () => {
     expect(1).toEqual(1)
   })
 
-  test('Test "toNotEqual()": 1 !== 2', () => {
-    expect(1).toNotEqual(2)
+  test('Test "not.toEqual()": 1 !== 2', () => {
+    expect(1).not.toEqual(2)
   })
 
   test('Test "toEqual()": () => {} === () => {}', () => {
     expect(() => {}).toEqual(() => {})
   })
 
-  test('Test "toNotEqual()": () => {} !== (parameter) => {}', () => {
-    expect(() => {}).toNotEqual((parameter) => {})
+  test('Test "not.toEqual()": () => {} !== (parameter) => {}', () => {
+    expect(() => {}).not.toEqual((parameter) => {})
   })
 
   test('Test "toEqual()": { a: 1 } === { a: 1 }', () => {
     expect({ a: 1 }).toEqual({ a: 1 })
   })
 
-  test('Test "toNotEqual()": { a: 1 } !== { a: 2 }', () => {
-    expect({ a: 1 }).toNotEqual({ a: 2 })
+  test('Test "not.toEqual()": { a: 1 } !== { a: 2 }', () => {
+    expect({ a: 1 }).not.toEqual({ a: 2 })
   })
 
   test('Test "toEqual()": [1] === [1]', () => {
     expect([1]).toEqual([1])
   })
 
-  test('Test "toNotEqual()": [1] !== ["a"]', () => {
-    expect([1]).toNotEqual(['a'])
+  test('Test "not.toEqual()": [1] !== ["a"]', () => {
+    expect([1]).not.toEqual(['a'])
   })
 
   test('Test "toEqual()": new Set([1]) === new Set([1])', () => {
     expect(new Set([1])).toEqual(new Set([1]))
   })
 
-  test('Test "toNotEqual()": new Set([1]) !== new Set(["a"])', () => {
-    expect(new Set([1])).toNotEqual(new Set(['a']))
+  test('Test "not.toEqual()": new Set([1]) !== new Set(["a"])', () => {
+    expect(new Set([1])).not.toEqual(new Set(['a']))
   })
 
   test("Test \"toEqual()\": new Map('a' [1]) === new Map('a' [1])", () => {
@@ -51,14 +51,14 @@ group('Equality tests', () => {
     expect(givenMap).toEqual(expectedMap)
   })
 
-  test("Test \"toNotEqual()\": new Map('a' [1]) !== new Map('a' [12])", () => {
+  test("Test \"not.toEqual()\": new Map('a' [1]) !== new Map('a' [12])", () => {
     const givenMap = new Map()
     givenMap.set('a', [1])
 
     const expectedMap = new Map()
     expectedMap.set('a', [12])
 
-    expect(givenMap).toNotEqual(expectedMap)
+    expect(givenMap).not.toEqual(expectedMap)
   })
 
   test('Test "toMatch()": " sdf" to "/sdf/"', () => {
@@ -73,14 +73,14 @@ group('Equality tests', () => {
     expect(' sdf').toMatch('sdf')
   })
 
-  test('Test "toNotMatch()": " sdf" to "/sdt/"', () => {
-    expect(' sdf').toNotMatch(/sdt/)
+  test('Test "not.toMatch()": " sdf" to "/sdt/"', () => {
+    expect(' sdf').not.toMatch(/sdt/)
   })
-  test('Test "toNotMatch()": " sdf" to "sdt"', () => {
-    expect(' sdf').toNotMatch('sdt')
+  test('Test "not.toMatch()": " sdf" to "sdt"', () => {
+    expect(' sdf').not.toMatch('sdt')
   })
 
-  test('Test "toNotMatch()": "sdf " to "/^[\\s]+sdf[\\s]+$/"', () => {
-    expect('sdf ').toNotMatch(/^[\s]+sdf$/)
+  test('Test "not.toMatch()": "sdf " to "/^[\\s]+sdf[\\s]+$/"', () => {
+    expect('sdf ').not.toMatch(/^[\s]+sdf$/)
   })
 })