+++ test/index.mjs
@@ -1,4 +1,5 @@
 import './value_type.mjs'
 import './equality.mjs'
 import './functions.mjs'
+import './coerce.mjs'
 import './promises.mjs'