+++ test/value_type.mjs
@@ -2,121 +2,121 @@ import { group, test, expect } from '../dist/index.mjs'
 
 group('Test methods, that exemine types', () => {
   test('1 is number', () => {
-    expect(1).isNumber()
+    expect(1).toBe('number')
   })
 
   test('"1" is not number', () => {
-    expect('1').isNotNumber()
+    expect('1').not.toBe('number')
   })
   test('NaN is NaN', () => {
-    expect(NaN).isNaN()
+    expect(NaN).toBe('NaN')
   })
 
   test('1 is not NaN', () => {
-    expect(1).isNotNaN()
+    expect(1).not.toBe('NaN')
   })
 
   test('"1" is string', () => {
-    expect('1').isString()
+    expect('1').toBe('string')
   })
 
   test('1 is not string', () => {
-    expect(1).isNotString()
+    expect(1).not.toBe('string')
   })
 
   test('true is boolean', () => {
-    expect(true).isBoolean()
+    expect(true).toBe('boolean')
   })
 
   test('"true" is not boolean', () => {
-    expect('true').isNotBoolean()
+    expect('true').not.toBe('boolean')
   })
 
   test('null is null', () => {
-    expect(null).isNull()
+    expect(null).toBe('null')
   })
 
   test('"null" is not null', () => {
-    expect('null').isNotNull()
+    expect('null').not.toBe('null')
   })
 
   test('undefined is undefined', () => {
-    expect(undefined).isUndefined()
+    expect(undefined).toBe('undefined')
   })
 
   test('"undefined" is not undefined', () => {
-    expect('undefined').isNotUndefined()
+    expect('undefined').not.toBe('undefined')
   })
 
   test('[] is Array', () => {
-    expect([]).isArray()
+    expect([]).toBe('Array')
   })
 
   test('{} is not Array', () => {
-    expect({}).isNotArray()
+    expect({}).not.toBe('Array')
   })
 
   test('Set is Set', () => {
-    expect(new Set()).isSet()
+    expect(new Set()).toBe('Set')
   })
 
   test('{} is not Set', () => {
-    expect({}).isNotSet()
+    expect({}).not.toBe('Set')
   })
 
   test('WeakSet is WeakSet', () => {
-    expect(new WeakSet()).isWeakSet()
+    expect(new WeakSet()).toBe('WeakSet')
   })
 
   test('Set is not WeakSet', () => {
-    expect(new Set()).isNotWeakSet()
+    expect(new Set()).not.toBe('WeakSet')
   })
 
   test('Map is Map', () => {
-    expect(new Map()).isMap()
+    expect(new Map()).toBe('Map')
   })
 
   test('{} is not Map', () => {
-    expect({}).isNotMap()
+    expect({}).not.toBe('Map')
   })
 
   test('WeakMap is WeakMap', () => {
-    expect(new WeakMap()).isWeakMap()
+    expect(new WeakMap()).toBe('WeakMap')
   })
 
   test('Map is not WeakMap', () => {
-    expect(new Map()).isNotWeakMap()
+    expect(new Map()).not.toBe('WeakMap')
   })
 
   test('function is function', () => {
-    expect(() => {}).isFunction()
+    expect(() => {}).toBe('function')
   })
 
   test('1 is not a function', () => {
-    expect(1).isNotFunction()
+    expect(1).not.toBe('function')
   })
 
   test('Promise is Promise', () => {
-    expect(new Promise((resolve, reject) => {})).isPromise()
+    expect(new Promise((resolve, reject) => {})).toBe('Promise')
   })
 
   test('1 is not a Promise', () => {
-    expect(1).isNotPromise()
+    expect(1).not.toBe('Promise')
   })
 
   test('/a/ is RegExp', () => {
-    expect(/a/).isRegExp()
+    expect(/a/).toBe('RegExp')
   })
 
   test('1 is not a RegExp', () => {
-    expect(1).isNotRegExp()
+    expect(1).not.toBe('RegExp')
   })
 
   test('{} is {}', () => {
-    expect({}).isPlainObject()
+    expect({}).toBe('PlainObject')
   })
 
   test('1 is not a {}', () => {
-    expect(1).isNotPlainObject()
+    expect(1).not.toBe('PlainObject')
   })
 })